Move fetch to @lodestar/api package
