+++ packages/api/src/index.ts
@@ -10,6 +10,9 @@ export {
   HttpError,
   ApiError,
   Metrics,
+  FetchError,
+  isFetchError,
+  fetch,
 } from "./utils/client/index.js";
 export * from "./utils/routes.js";
