+++ packages/api/src/utils/client/index.ts
@@ -1,2 +1,3 @@
 export * from "./client.js";
 export * from "./httpClient.js";
+export * from "./fetch.js";
