+++ packages/api/test/utils/fetchOpenApiSpec.ts
@@ -1,6 +1,6 @@
 import fs from "node:fs";
 import path from "node:path";
-import {fetch} from "@lodestar/utils";
+import {fetch} from "@lodestar/api";
 import {OpenApiFile, OpenApiJson} from "./parseOpenApiSpec.js";
 
 /* eslint-disable no-console */
