+++ packages/beacon-node/src/execution/engine/utils.ts
@@ -1,4 +1,4 @@
-import {isFetchError} from "@lodestar/utils";
+import {isFetchError} from "@lodestar/api";
 import {IJson, RpcPayload} from "../../eth1/interface.js";
 import {IJsonRpcHttpClient} from "../../eth1/provider/jsonRpcHttpClient.js";
 import {ExecutePayloadStatus, ExecutionEngineState} from "./interface.js";
