+++ packages/beacon-node/test/unit/metrics/server/http.test.ts
@@ -1,4 +1,4 @@
-import {fetch} from "@lodestar/utils";
+import {fetch} from "@lodestar/api";
 import {getHttpMetricsServer, HttpMetricsServer} from "../../../../src/metrics/index.js";
 import {testLogger} from "../../../utils/logger.js";
 import {createMetricsTest} from "../utils.js";
