+++ packages/validator/src/util/externalSignerClient.ts
@@ -1,7 +1,7 @@
 import {ContainerType, toHexString, ValueOf} from "@chainsafe/ssz";
+import {fetch} from "@lodestar/api";
 import {phase0, altair, capella} from "@lodestar/types";
 import {ForkSeq} from "@lodestar/params";
-import {fetch} from "@lodestar/utils";
 import {ValidatorRegistrationV1} from "@lodestar/types/bellatrix";
 import {BeaconConfig} from "@lodestar/config";
 import {computeEpochAtSlot, blindedOrFullBlockToHeader} from "@lodestar/state-transition";
